+++ lib/emailbutler/webhooks/receiver.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'emailbutler/webhooks/mappers/sendgrid'
+
 module Emailbutler
   module Webhooks
     class Receiver
@@ -10,21 +12,21 @@ module Emailbutler
       end
 
       def call(user_agent:, payload:)
-        case user_agent
-        when SENDGRID_USER_AGENT then sengrid_message(payload)
-        end
+        select_mapper(user_agent)
+          .call(payload: payload)
+          .each { |message|
+            Emailbutler.update_message(
+              Emailbutler.find_message_by(uuid: message[:uuid]),
+              status: message[:status]
+            )
+          }
       end
 
       private
 
-      def sengrid_message(payload)
-        payload['_json'].each do |event|
-          next unless event['smtp-id']
-
-          message = Emailbutler.find_message_by(uuid: event['smtp-id'])
-          next unless message
-
-          Emailbutler.update_message(message, status: event['event'])
+      def select_mapper(user_agent)
+        case user_agent
+        when SENDGRID_USER_AGENT then Emailbutler::Webhooks::Mappers::Sendgrid
         end
       end
     end